Add phase-inversion to bh_pan and remove inverse_phase plugin

+++ libblankenhain/src/Filter.cpp
@@ -57,7 +57,7 @@ void Filter<T>::setBell(double frequency, double Q, double gainInDb)
 template <typename T>
 void Filter<T>::setLowShelf(double frequency, double Q, double gainInDb)
 {
-	const float A = aux::decibelToLinear(gainInDb * 0.5f);
+	const float A = aux::decibelToLinear(static_cast<float>(gainInDb) * 0.5f);
 	recomputeCoefficients(frequency, Q, 1. / BhMath::sqrt(static_cast<float>(A)));
 	m0 = T(1);
 	m1 = k * T(A - 1);
@@ -67,7 +67,7 @@ void Filter<T>::setLowShelf(double frequency, double Q, double gainInDb)
 template <typename T>
 void Filter<T>::setHighShelf(double frequency, double Q, double gainInDb)
 {
-	const float A = aux::decibelToLinear(gainInDb * 0.5f);
+	const float A = aux::decibelToLinear(static_cast<float>(gainInDb) * 0.5f);
 	recomputeCoefficients(frequency, Q, BhMath::sqrt(static_cast<float>(A)));
 	m0 = T(A * A);
 	m1 = k * T((1 - A) * A);
